Added text content of Learn section
A 2012 University of Michigan study showed that mindfulness yoga, which combines physical poses with meditation practices, can bring measurable relief to the depression that can accompany the emotional journey of pregnancy. Of course, not all symptoms are guaranteed to disappear altogether, but the multidimensional approach of yoga to both physical and emotional health can help your body take the uncomfortable aspects of pregnancy in stride.